+++ conquest2a/conquest.py
@@ -12,6 +12,25 @@ import conquest2a._types as c2at
 
 @dataclass
 class Atom:
+    """Class which holds Atom data.
+
+    :param species: The integer referring to the species as defined in the ``Conquest_input`` file.
+    :type species: ``int``
+    :param coords: Fractional coordinates of the atom in the interval :math:`[0,1)`. Numbers outside this range are wrapped back into the range.
+    :type coords: :ref:`REAL ARRAY <types>`
+    :param can_move: Whether the atom is allowed to move in each Cartesian axis.
+    :type can_move: ``Sequence[str]`` - list of 'T' or 'F'.
+    :param number: The atom number, i.e. its position in the coordinates file.
+    :type number: ``int``
+    :param label: The atom element
+    :type label: ``str``
+    :param cart_coords: Cartesian coordinates of the atom.
+    :type cart_coords: :ref:`REAL ARRAY <types>`
+    :param forces: Force vector of the atom, defaults to ``np.array([0.0, 0.0, 0.0])``.
+    :type forces: :ref:`REAL ARRAY <types>`
+    :param spins: Spin moment on the atom,  defaults to ``np.array([0.0, 0.0, 0.0])``.
+    :type spins: :ref:`REAL ARRAY <types>`
+    """
     species: int
     coords: c2at.REAL_ARRAY
     can_move: Sequence[str]
@@ -39,14 +58,15 @@ class Atom:
 
 class conquest_input:
     def __init__(self, species_dict: dict[int, str]) -> None:
-        """Constructor for conquest_input
+        """`conquest_input` serves as the main entrypoint describing the species involved in the simulation.
 
-        Args:
-            species_dict (dict[int, str]): hashmap species_index <-> element_label
+        :param species_dict: hashmap species_index <-> element_label
                 It is not completely reliable to directly read conquest_input
                 E.g., for multiple spins, must duplicate an element and call it a
                 new Conquest species, however labels can be any alphanumeric string
                 Therefore we expect a dictionary to be passed in independently.
+        :type species_dict: dict[int, str]
+        :raises ValueError: If incorrect elements are passed in, the code will abort.
         """
         self.element_file = "elements.txt"
         self.species_dict = species_dict
@@ -66,6 +86,13 @@ class conquest_input:
 
 class processor_base:
     def __init__(self, path: str, err_str: str | None = None) -> None:
+        """Generic parent class for any processor classes
+
+        :param path: Path to a file
+        :type path: ``str``
+        :param err_str: An error message to print if something goes wrong, defaults to None
+        :type err_str: ``str | None``, optional
+        """
         self.input_path = path.strip()
         self.abs_input_path: Path
         self.err_str = err_str
@@ -73,6 +100,13 @@ class processor_base:
         self.re_index = re.compile(r"\d+")
 
     def resolve_path(self, filename: str | None = None) -> None:
+        """_summary_
+
+        :param filename: Path to a file, defaults to None, which will use the ``path`` passed into the class instance.
+        :type filename: ``str | None``, optional
+        :raises FileNotFoundError: Exits if the file is not found.
+        :raises RuntimeError: Will exit if the file is corrupted
+        """
         abs_coord_path: Path = Path("")
         if filename is None:
             abs_coord_path = Path(os.path.abspath(self.input_path))
@@ -92,19 +126,29 @@ class processor_base:
 
 
 class conquest_coordinates:
+    """Class which holds data about the system, including the :class:`Atom` s in the system.
+    
+    :param conquest_input: :class:`conquest_input` instance.
+    :type conquest_input: ``conquest_input``
+    """
     def __init__(
         self,
         conquest_input: conquest_input,
     ) -> None:
+        
         self.atoms: list[Atom] = []
         self.conquest_input = conquest_input
         self.natoms: str
         self.element_map: dict[str, list[Atom]]
         self.lattice_vectors: c2at.REAL_ARRAY = np.array([])
-
         self.cart_position_vectors: c2at.REAL_ARRAY = np.array([])
 
     def get_cartesian_positions(self) -> c2at.REAL_ARRAY:
+        """Returns the Cartesian position of all the atoms in the system and attaches to its :class:`Atom` field.
+
+        :return: The 3D vector of the Cartesian position.
+        :rtype: :ref:`REAL ARRAY <types>`
+        """
         atom_frac_pos = np.vstack([atom.coords for atom in self.atoms])
         cart_coords = atom_frac_pos @ self.lattice_vectors.T
         for atom, cart_coord in zip(self.atoms, cart_coords):
@@ -113,14 +157,14 @@ class conquest_coordinates:
         return cart_coords
 
     def assign_atom_labels(self) -> None:
-        """Assign each Atom its label"""
+        """Assign each Atom its label.
+        If the ``conquest_input`` does not define labels for all species, these species labels will be silently skipped.
+        """
         for atom in self.atoms:
             atom.label = self.conquest_input.species_dict[atom.species]
 
     def index_to_atom_map(self) -> None:
-        """Every Atom now has its element label.
-        External file formats require a count of the number of Atoms per element,
-        so we now form a dict of elements to Atoms in preparation for writing"""
+        """Form a dictionary with keys an element label, and values a list of all the :class:`Atom` s with that label. External file formats, such as ``.vasp``, require a count of the number of atoms per element."""
         ele_to_atom: dict[str, list[Atom]] = {}
         for element in self.conquest_input.unique_elements:
             # print(list(a for a in self.Atoms if a.label == element))
@@ -128,6 +172,11 @@ class conquest_coordinates:
         self.element_map = ele_to_atom
 
     def number_of_elements(self) -> dict[str, int]:
+        """Function to get the number of atoms of each element. 
+
+        :return: Returns a dictionary of the number of atoms per element
+        :rtype: ``dict[str, int]``
+        """
         num_eles: dict[str, int] = {}
         for element in list(self.element_map.keys()):
             num_eles[element] = len(self.element_map[element])
@@ -135,6 +184,13 @@ class conquest_coordinates:
 
 
 class conquest_coordinates_processor(processor_base):
+    """Class which extracts data from a CONQUEST coordinates file and populates a :class:`conquest_coordinates` instance.
+
+        :param path: Path of the CONQUEST coordinates file to read.
+        :type path: ``str``
+        :param conquest_input: :class:`conquest_input` instance.
+        :type conquest_input: conquest_input
+    """
     def __init__(self, path: str, conquest_input: conquest_input) -> None:
         processor_base.__init__(
             self, path=path, err_str="Error opening specified CONQUEST coordinates file."
@@ -152,12 +208,13 @@ class conquest_coordinates_processor(processor_base):
         self.volume_ang = self.volume_bohr * BOHR_TO_ANGSTROM_VOLUME
 
     def open_file(self) -> None:
-        """
+        """This method reads a CONQUEST coordinate file.
+
         CONQUEST coords file split into 3 main chunks:
-        first 3 lines are lattice vectors
-        fourth line is the total number of atoms in the unit cell
-        the following lines are of the for_summary_m:
-          <double> <double> <double> <int> <char> <char> <char>
+            * first 3 lines are lattice vectors
+            * fourth line is the total number of atoms in the unit cell
+            * the following lines describe each atom and look like
+                <double> <double> <double> <int> <char> <char> <char>
         """
         with open(self.abs_input_path, "r", encoding="utf-8") as conquest_coord_file:
             conquest_lattice_data_str = [next(conquest_coord_file).strip() for _ in range(3)]
@@ -193,6 +250,11 @@ class atom_charge(processor_base):
     same CONQUEST input coordinates file.
 
     In particular, make use of the conquest_coordinates class to contain the list of Atoms
+    
+    :param coordinates: The :class:`conquest_coordinates` instance to use
+    :type coordinates: :class:`conquest_coordinates`
+    :param atom_charge_path: Path to the ``AtomCharge.dat`` file.
+    :type atom_charge_path: ``str``
     """
 
     def __init__(self, coordinates: conquest_coordinates, atom_charge_path: str) -> None:
@@ -205,19 +267,14 @@ class atom_charge(processor_base):
         self.atom_charge_path = atom_charge_path
         self.abs_atom_charge_path: Path = Path(os.path.abspath(self.atom_charge_path))
         self.conquest_charge_data: list[c2at.REAL_ARRAY] = []
-
-        try:
-            self.resolve_path()
-            self.open_file()
-        except FileNotFoundError as e:
-            print(e)
+        self.resolve_path()
+        self.open_file()
         self.assign_atom_charge()
 
     def open_file(self) -> None:
-        """
-        CONQUEST AtomCharge.dat
-          <double> <double> <double>
-          total, "up", "down" - CONQUEST only deals with collinear spins
+        """Reads in the AtomCharge.dat
+        Format of the file is 3 columns: total, spin up spin down
+        CONQUEST only deals with collinear spins
         """
         with open(self.abs_atom_charge_path, "r", encoding="utf-8") as conquest_charge_file:
             for line in conquest_charge_file:
@@ -239,10 +296,10 @@ class atom_charge(processor_base):
 
 
 class block_processor:
-    """Generic class to process CONQUEST output files split into blocks via &
+    """Generic class to process CONQUEST output files which are split into blocks via `&`
 
     Sometimes these blocks are categorised by spins and have comments
-    # at the start of the category
+    # at the start of the section
     """
 
     def __init__(self) -> None:
@@ -254,15 +311,20 @@ class block_processor:
 
     def process_headers(self, line: str) -> None:
         """This function can be overridden by subclasses to process header lines starting with #
-        Bandstructure and DOS/pDOS files have slightly different headers
+        Bandstructure and DOS/pDOS files have slightly different headers.
         """
         pass
 
     def process_block(self, line: str) -> None:
-        """In CONQUEST, blocks are separated by & on its own newline"""
+        """Extracts blocks, which are separated by ``&`` on its own newline, for data processing."""
         pass
 
     def read_file(self, filename: str) -> None:
+        """Generic method which calls :func:`~process_headers` and :func:`~process_block`.
+
+        :param filename: Path to file
+        :type filename: ``str``
+        """
         self.blocks = []
         with open(filename, "r", encoding="utf-8") as f:
             for line in f: