+++ conquest2a/writers.py
@@ -7,6 +7,17 @@ from conquest2a.constants import BOHR_TO_ANGSTROM
 
 
 class file_writer:
+    """Generic parent class to define file operations and variables.
+
+        :param dest: _description_
+        :type dest: str
+        :param mode: _description_, defaults to "w"
+        :type mode: str, optional
+        :param encoding: _description_, defaults to "utf-8"
+        :type encoding: str, optional
+        :param is_angstrom: _description_, defaults to False
+        :type is_angstrom: bool, optional
+        """
     def __init__(
         self, dest: str, mode: str = "w", encoding: str = "utf-8", is_angstrom: bool = False
     ) -> None:
@@ -28,6 +39,18 @@ class file_writer:
 
 
 class conquest_writer(file_writer):
+    """Class to write a CONQUEST coordinates file given a :class:`conquest_coordinates` instance.
+
+        :param dest: Path to write the new coordinates file.
+        :type dest: ``str``
+        :param coords: :class:`conquest_coordinates` instance to write.
+        :type coords: ``conquest_coordinates``
+        :param encoding: File encoding, defaults to "utf-8"
+        :type encoding: ``str``, optional
+        :param precision: Float precision, defaults to 10
+        :type precision: ``int``, optional
+        :raises ValueError: If the float precision is not at least 1.
+        """
     def __init__(
         self,
         dest: str,
@@ -66,6 +89,17 @@ class conquest_writer(file_writer):
 
 
 class vasp_writer(file_writer):
+    """Class to write a VASP file given a :class:`~conquest.conquest_coordinates` instance.
+
+        :param dest: Path to write the new coordinates file.
+        :type dest: ``str``
+        :param data: :class:`~conquest.conquest_coordinates` instance to write.
+        :type data: ``conquest_coordinates``
+        :param encoding: File encoding, defaults to "utf-8"
+        :type encoding: ``str``, optional
+        :param is_angstrom: Whether the data in ``conquest_coordinates`` is already in angstroms instead of Bohrs, defaults to ``False``.
+        :type is_angstrom: ``bool``, optional
+        """
     def __init__(
         self,
         dest: str,
@@ -89,9 +123,14 @@ class vasp_writer(file_writer):
         with self.file as file:
             file.write(f"{ele_string}\n")
             file.write("1.0\n")
-            for lattice_vect in self.data.lattice_vectors:
-                file.write(rf'  {" ".join(str(x * BOHR_TO_ANGSTROM) for x in lattice_vect)}')
-                file.write("\n")
+            if self.is_ang:
+                for lattice_vect in self.data.lattice_vectors:
+                    file.write(rf'  {" ".join(str(x) for x in lattice_vect)}')
+                    file.write("\n")
+            else:
+                for lattice_vect in self.data.lattice_vectors:
+                    file.write(rf'  {" ".join(str(x * BOHR_TO_ANGSTROM) for x in lattice_vect)}')
+                    file.write("\n")
             file.write(f"{ele_string}\n")
             file.write(f"{num_string}\n")
             file.write("Direct\n")
@@ -102,7 +141,17 @@ class vasp_writer(file_writer):
 
 
 class xyz_writer(file_writer):
-    """Class to generate .xyz for basic XYZ file format."""
+    """Class to write a ``.xyz`` for a basic XYZ file given a :class:`~conquest.conquest_coordinates` instance.
+
+        :param dest: Path to write the new coordinates file.
+        :type dest: ``str``
+        :param data: :class:`~conquest.conquest_coordinates` instance to write.
+        :type data: ``conquest_coordinates``
+        :param encoding: File encoding, defaults to "utf-8"
+        :type encoding: ``str``, optional
+        :param comment_line: What string to write as the comment line
+        :type comment_line: ``str``, optional
+        """
 
     def __init__(
         self,
@@ -121,6 +170,11 @@ class xyz_writer(file_writer):
         return self.comment_line
 
     def create_atoms_str(self) -> tuple[str, str]:
+        """Creates a line for an atom, for the XYZ file.
+
+        :return: Components of the line to write: the element and the numbers
+        :rtype: tuple[str, str]
+        """
         num_ele = self.data.number_of_elements()
         ele_string = " ".join(list(num_ele.keys()))
         num_string = " ".join(str(x) for x in num_ele.values())
@@ -144,6 +198,19 @@ class xyz_writer(file_writer):
 
 
 class extxyz_writer(xyz_writer):
+    """Class to write a ``.extxyz`` for a basic XYZ file given a :class:`~conquest.conquest_coordinates` instance.
+
+        The main advantage of `.extxyz` is the ability to specify columns and the time, which is very useful for animations. I recommend just using CONQUEST's ability to output ``.extxyz`` files at different timesteps.
+
+        :param dest: Path to write the new coordinates file.
+        :type dest: ``str``
+        :param data: :class:`~conquest.conquest_coordinates` instance to write.
+        :type data: ``conquest_coordinates``
+        :param encoding: File encoding, defaults to "utf-8".
+        :type encoding: ``str``, optional
+        :param time: The instance of time of the cell, defaults to `0.0`.
+        :type time: ``float``, optional
+        """
     def __init__(
         self,
         dest: str,
@@ -151,10 +218,16 @@ class extxyz_writer(xyz_writer):
         encoding: str = "utf-8",
         time: float = 0.0,
     ) -> None:
+        
         self.time = time
         super().__init__(dest=dest, data=data, encoding=encoding)
 
     def create_comment_line(self) -> str:
+        """Creates the ``.extxyz`` comment line: specifies columns, formats and time.
+
+        :return: The file's comment line.
+        :rtype: ``str``
+        """
         lattice: list[float] = []
         for single_vector in self.data.lattice_vectors:
             lattice.append(single_vector[0])
@@ -167,6 +240,19 @@ class extxyz_writer(xyz_writer):
 
 
 class xsf_writer(file_writer):
+    """Class to write `.xsf` files given a :class:`~conquest.conquest_coordinates` instance.
+
+        XSF files support an extra 3 columns alongside the 3 columns used for position. These columns specify a vector associated with each atom. In CONQUEST, the relevant vectors are forces and spins. Note that CONQUEST only supports *collinear spin* (up and down), so the spin vector is visualised along the :math:`c`-axis of the simulation cell.
+
+        :param dest: Path to write the new coordinates file.
+        :type dest: ``str``
+        :param data: :class:`~conquest.conquest_coordinates` instance to write.
+        :type data: ``conquest_coordinates``
+        :param encoding: File encoding, defaults to "utf-8".
+        :type encoding: ``str``, optional
+        :param write_extra: Whether to extract the force vector or spin vector of atoms, defaults to "spin".
+        :type write_extra: Literal["spin", "force"], optional
+        """
     def __init__(
         self,
         dest: str,
@@ -205,13 +291,24 @@ class xsf_writer(file_writer):
 
 
 class xsf_writer_spins(file_writer):
+    """
+        XSF writer class that includes spin information from the AtomCharge.dat file,
+        by editing an existing XSF file generated by PostProcessCQ.
+
+        :param dest: Path to write the new coordinates file.
+        :type dest: ``str``
+        :param data: :class:`~conquest.conquest_coordinates` instance to write.
+        :type data: ``conquest_coordinates``
+        :param encoding: File encoding, defaults to "utf-8".
+        :type encoding: ``str``, optional
+        :param charges: :class:`~conquest.atom_charge` instance to read spin data from.
+        :type charges: :class:`~conquest.atom_charge`
+         :param xsf_file: XSF file to modify.
+        :type xsf_file: ``str``
+        """
     def __init__(
         self, dest: str, charges: atom_charge, xsf_file: str, encoding: str = "utf-8"
     ) -> None:
-        """
-        XSF writer class that includes spin information from the AtomCharge.dat file,
-        by editing an existing XSF file generated by PostProcessCQ.
-        """
         self.charges = charges
         self.original_xsf_file = xsf_file.strip()
         self.encoding = encoding